Tried to install VST without GUI
I got tired of having to press trough the Poise VST installer, so I
tried to load it through a configuration file. It doesn't seem to work
though, and the actual configuration file seems to confirm that since it
doesn't actually store the most important data.
It does fill out the options correctly, but that might be because it
finds the existing installation in the wine prefix.