Made things very close to working.
I removed the errors. And since I couldn't for the life of me figure out
how to properly debug it (builtins.trace sucks since it won't evaluate
the property, and won't do it recursively) I just winged it and
guessed another way to do it. And it worked somehow. Looking over the
generated files, this might very well start working in the next commit!