Removed runtimeInputs error, now it does nothing
Finally fixed the runtimeInputs that caused me so much debugging
headache. I replaced "+" with "++".
That's it. That's all that was needed. When it was a single "+" nix
thought that I wanted string concatenation but failed as they weren't
string but lists.

Also, I realized that since my fish script generator doesn't pass the
bash PATH variable down, runtimeInputs still don't actually do anyting
even if they